Finalizando correcciones del primer comentario
