+++ src/Redux/User/Actions/userActions.js
@@ -1,0 +1,6 @@
+export const saveUser = userName => ({
+    type: 'SAVE_USERNAME',
+    payload: {
+      user: userName,
+    },
+  });
