+++ src/Redux/User/Reducers/userReducer.js
@@ -1,0 +1,18 @@
+const initalState = {
+    userName:'',
+}
+
+const userReducer = (state = initalState, action) => {
+    switch(action.type){
+       case 'SAVE_USERNAME':
+       return {
+          ...state,  
+          userName: action.payload.user,
+        };
+
+       default:
+       return state;
+    }
+  }; 
+
+export default userReducer
