+++ src/Redux/rootReducer.js
@@ -1,12 +1,13 @@
-import operationReducer from './operationReducer'
+import operationReducer from './Operations/Reducers/operationReducer'
+import userReducer from './User/Reducers/userReducer'
 import { reducer as formReducer } from 'redux-form'
 
 const {combineReducers} = require('redux')
 
 const rootReducer = combineReducers({
-    opRed: operationReducer, 
+    operation: operationReducer,
+    user:userReducer, 
     form: formReducer,
-    
 })
 
 export default rootReducer
